pyproject.toml
[project]
name = "dm-recommender"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.11"
dependencies = [
    "beautifulsoup4>=4.14.2",
    "psycopg2-binary>=2.9.11",
    "requests>=2.32.5",
]
